Production configuration

+++ src/modules/users/services/UpdateUserAvatarService.ts
@@ -8,6 +8,7 @@ import IStorageProvider from '@shared/container/providers/StorageProvider/models
 import AppError from '@shared/errors/AppError';
 
 import User from '../infra/typeorm/entities/User';
+import ICacheProvider from '@shared/container/providers/CacheProvider/models/ICacheProvider';
 
 interface IRequest {
   user_id: string;
@@ -21,7 +22,10 @@ class UpdateUserAvatarService {
     private usersRepository: IUsersRepository,
 
     @inject('StorageProvider')
-    private storageProvider: IStorageProvider
+    private storageProvider: IStorageProvider,
+
+    @inject('CacheProvider')
+    private cacheProvider: ICacheProvider,
   ) { }
 
   public async execute({ user_id, avatarFilename }: IRequest): Promise<User> {
@@ -31,7 +35,7 @@ class UpdateUserAvatarService {
       throw new AppError('Only authenticated users can change avatar.', 401);
     }
 
-    if (user.avatar) {
+    if (user.avatar && user.avatar !== 'default.png') {
       await this.storageProvider.deleteFile(user.avatar);
     }
 
@@ -40,6 +44,8 @@ class UpdateUserAvatarService {
     user.avatar = fileName;
 
     await this.usersRepository.save(user);
+
+    await this.cacheProvider.invalidatePrefix('providers-list');
 
     return user;
   }
